requirements.txt
google-generativeai==0.8.4
tweepy==4.14.0
pandas==2.2.0
streamlit==1.32.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
matplotlib==3.8.0
